Generate pundit:install. Generated pundit:policy for PostIt with the following rules:   - scope: only current user's postits   - create: no condition (true)   - edit: only current user's postits   - update: only current user's postits   - destroy: only current user's postits Deleted PostIts#new due to inutility.

+++ app/controllers/postits_controller.rb
@@ -4,8 +4,9 @@ class PostitsController < ApplicationController
   # Define a colors constant
   COLORS = ["#86A1DA", "#88D6BA", "#F23595", "#283139", "#FEE269"]
 
-  def new
+  def index
     @postit = Postit.new
+    @postits = policy_scope(Postit)
   end
 
   def create
@@ -15,7 +16,8 @@ class PostitsController < ApplicationController
     @postit.color = COLORS.sample
     # Set Postit user to current user
     @postit.user = current_user
-
+    # Check if user authorized
+    authorize @postit
     # Save and check if successful
     if @postit.save
       # Redirect to postits#index
@@ -27,10 +29,14 @@ class PostitsController < ApplicationController
 
   def edit
     # Set post using id from params (before_action :set_postit)
+    # Check if user authorized
+    authorize @postit
   end
 
   def update
     # Set post using id from params (before_action :set_postit)
+    # Check if user authorized
+    authorize @postit
     # Update and check if successful
     if @postit.update(postit_params)
       # Redirect to postits#index
@@ -42,6 +48,8 @@ class PostitsController < ApplicationController
 
   def destroy
     # Set post using id from params (before_action :set_postit)
+    # Check if user authorized
+    authorize @postit
     # Destroy and check if successful
     if @postit.destroy
       # Redirect to postits#index
@@ -49,11 +57,6 @@ class PostitsController < ApplicationController
       # Flash with a nice message to confirm action success
       flash[:alert] = "Postit destroyed Modafucka!"
     end
-  end
-
-  def index
-    @postit = Postit.new
-    @postits = Postit.all
   end
 
   private
